+++ CHANGELOG.md
@@ -5,6 +5,17 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Adds `smtp` to receive and save SMTP event.
+- Adds SMTP to GraphQL API
+
+### Changed
+
+- Adds `answer` field of DNS event.
+
 ## [0.4.0] - 2022-11-01
 
 ### Added